requirements.txt
numpy>=1.21.0
pandas>=1.5.0
scipy>=1.9.0
scikit-learn>=1.1.0
yfinance>=0.2.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0
requests>=2.25.0
beautifulsoup4>=4.11.0
matplotlib>=3.5.0
seaborn>=0.11.0
xgboost>=1.6.0
lightgbm>=3.3.0
joblib>=1.1.0
python-dotenv>=0.19.0
schedule>=1.1.0
websocket-client>=1.3.0
